Cleaning up comments
