+++ api/bolt/migrator/migrator.go
@@ -10,7 +10,6 @@ import (
 	"github.com/portainer/portainer/bolt/stack"
 	"github.com/portainer/portainer/bolt/user"
 	"github.com/portainer/portainer/bolt/version"
-	"github.com/portainer/portainer/bolt/webhooks"
 )
 
 type (
@@ -25,7 +24,6 @@ type (
 		stackService           *stack.Service
 		userService            *user.Service
 		versionService         *version.Service
-		webhookService         *webhooks.Service
 		fileService            portainer.FileService
 	}
 
@@ -40,7 +38,6 @@ type (
 		StackService           *stack.Service
 		UserService            *user.Service
 		VersionService         *version.Service
-		WebhookService         *webhooks.Service
 		FileService            portainer.FileService
 	}
 )
@@ -57,7 +54,6 @@ func NewMigrator(parameters *Parameters) *Migrator {
 		stackService:           parameters.StackService,
 		userService:            parameters.UserService,
 		versionService:         parameters.VersionService,
-		webhookService:         parameters.WebhookService,
 		fileService:            parameters.FileService,
 	}
 }
